+++ src/main/java/dev/zprestige/ruby/module/combat/FeetPlace.java
@@ -3,10 +3,10 @@ package dev.zprestige.ruby.module.combat;
 import dev.zprestige.ruby.eventbus.annotation.RegisterListener;
 import dev.zprestige.ruby.events.PacketEvent;
 import dev.zprestige.ruby.module.Module;
-import dev.zprestige.ruby.newsettings.impl.ComboBox;
-import dev.zprestige.ruby.newsettings.impl.Parent;
-import dev.zprestige.ruby.newsettings.impl.Slider;
-import dev.zprestige.ruby.newsettings.impl.Switch;
+import dev.zprestige.ruby.settings.impl.ComboBox;
+import dev.zprestige.ruby.settings.impl.Parent;
+import dev.zprestige.ruby.settings.impl.Slider;
+import dev.zprestige.ruby.settings.impl.Switch;
 import dev.zprestige.ruby.util.BlockUtil;
 import dev.zprestige.ruby.util.EntityUtil;
 import dev.zprestige.ruby.util.InventoryUtil;
@@ -88,7 +88,7 @@ public class FeetPlace extends Module {
         }
         if (onMoveCancel.GetSwitch() && EntityUtil.isMoving())
             return;
-        if (!placeTimer.getTime(placeDelay.GetSlider()))
+        if (!placeTimer.getTime((long) placeDelay.GetSlider()))
             return;
         int slot = -1;
         switch (block.GetCombo()) {
