+++ src/main/java/dev/zprestige/ruby/module/misc/InventoryCleaner.java
@@ -1,9 +1,9 @@
 package dev.zprestige.ruby.module.misc;
 
 import dev.zprestige.ruby.module.Module;
-import dev.zprestige.ruby.newsettings.impl.Parent;
-import dev.zprestige.ruby.newsettings.impl.Slider;
-import dev.zprestige.ruby.newsettings.impl.Switch;
+import dev.zprestige.ruby.settings.impl.Parent;
+import dev.zprestige.ruby.settings.impl.Slider;
+import dev.zprestige.ruby.settings.impl.Switch;
 import dev.zprestige.ruby.util.InventoryUtil;
 import dev.zprestige.ruby.util.Timer;
 import net.minecraft.init.Blocks;
@@ -25,7 +25,7 @@ public class InventoryCleaner extends Module {
 
     @Override
     public void onTick() {
-        if (throwTimer.getTime(throwDelay.GetSlider()) && getItemSlot() != -1)
+        if (throwTimer.getTime((long) throwDelay.GetSlider()) && getItemSlot() != -1)
             throwAway(getItemSlot());
     }
