+++ src/main/java/dev/zprestige/ruby/module/player/AutoArmor.java
@@ -1,8 +1,7 @@
 package dev.zprestige.ruby.module.player;
 
 import dev.zprestige.ruby.module.Module;
-import dev.zprestige.ruby.newsettings.impl.Slider;
-import dev.zprestige.ruby.setting.impl.IntegerSetting;
+import dev.zprestige.ruby.settings.impl.Slider;
 import dev.zprestige.ruby.util.InventoryUtil;
 import dev.zprestige.ruby.util.Timer;
 import net.minecraft.init.Items;
@@ -21,7 +20,7 @@ public class AutoArmor extends Module {
     public void onTick() {
         if (mc.currentScreen != null)
             return;
-        if (getSlot() != -1 && timer.getTime(takeOnDelay.GetSlider())) {
+        if (getSlot() != -1 && timer.getTime((long) takeOnDelay.GetSlider())) {
             mc.playerController.windowClick(mc.player.inventoryContainer.windowId, getSlot(), 0, ClickType.QUICK_MOVE, mc.player);
             timer.setTime(0);
         }
